requirements.txt
# Core Django
Django==4.2.9
asgiref==3.11.1
sqlparse==0.5.5

# REST API
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
PyJWT==2.11.0
drf-spectacular==0.29.0

# Django extensions
django-filter==23.5
django-cors-headers==4.3.1
python-decouple==3.8

# File & Image
Pillow==10.2.0
openpyxl==3.1.2
et_xmlfile==2.0.0

# PDF
reportlab==4.1.0
PyMuPDF==1.27.1
pypdf==6.7.5
pypdfium2==5.6.0
pdfplumber==0.11.9
pdfminer.six==20251230

# Document
python-docx==1.2.0
lxml==6.0.2

# OpenAPI / Schema (dependencies drf-spectacular)
PyYAML==6.0.3
uritemplate==4.2.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
referencing==0.37.0
rpds-py==0.30.0
attrs==25.4.0

# Crypto & Security
cryptography==46.0.5
cffi==2.0.0
pycparser==3.0

# Utils
typing_extensions==4.15.0
chardet==7.0.1
charset-normalizer==3.4.5
pytz==2026.1.post1
tzdata==2025.3
# Deploy
dj-database-url==2.1.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
